Options test project renamed
