+++ OgcApi.Options.Tests/JsonDeserializationFacts.cs
@@ -1,17 +1,15 @@
 using OgcApi.Net.Options.Features;
 using Xunit;
 
-namespace OgcApi.Net.Features.Tests
+namespace OgcApi.Options.Tests
 {
-    public class ConfigurationDeserializationFacts : IClassFixture<ConfigurationOptionsFixture>
+    public class JsonDeserializationFacts : IClassFixture<JsonOptionsFixture>
     {
-        private readonly ConfigurationOptionsFixture _fixture;
-
-        public ConfigurationDeserializationFacts(ConfigurationOptionsFixture fixture)
+        private readonly JsonOptionsFixture _fixture;
+        public JsonDeserializationFacts(JsonOptionsFixture fixture)
         {
             _fixture = fixture;
         }
-
         [Fact]
         public void OgcApiOptionsDeserialization()
         {
@@ -261,6 +259,7 @@ namespace OgcApi.Net.Features.Tests
             Assert.Equal(4, extent.Temporal.Interval[1][1]);
         }
     }
+
 }
