+++ OgcApi.Options.Tests/ProvidersOptionsFixture.cs
@@ -1,8 +1,8 @@
 ﻿using OgcApi.Net.DataProviders;
-using OgcApi.Net.Features.Tests.Utils;
+using OgcApi.Options.Tests.Utils;
 using OgcApi.Net.Options;
 
-namespace OgcApi.Net.Features.Tests
+namespace OgcApi.Options.Tests
 {
 
     public class ProvidersOptionsFixture
